Fix recompilation (potential) issue
Because module attributes requires to recompile the deps to be
able to pick up the change. It is better to not use it for
configuration.

If somebody compiles the swoosh package with Jason and later on
decides to change it to Poison it will require to recompile the
swoosh deps.

Elixir do not recompile the deps unless it detects a change and
since change the config value is not change it will still have
the old value of the configuration.