Short Circuit Delivery Many
Sometimes I'll do this:

```elixir
some_list
|> Enum.map(&to_swoosh_email)
|> Swoosh.Adapters.Mailjet.deliver_many()
```

and sometimes some_list will be an empty list. Currently if it is empty
I will get an error from mailjet, when really it isn't an error, I just
have to emails to send.

This short circuits the email send when there are no emails to send.