Fix crashes caused by the memory GenServer restarts
Connecting/disconnecting nodes to a cluster will cause multiple restarts of the `Memory` GenServer making exceeding the `max_restarts` of the top-level supervisor easy thus crashing the application.

This adds a manager GenServer that will monitor the global Memory one instead of linking to it to avoid crashing it when a node disconnects and trigger the above condition.

The bug is triggered by having two nodes in a cluster and running this code:
```
def trigger do
    Node.connect(:"node1@127.0.0.1")
    Process.sleep(400)
    Node.disconnect(:"node1@127.0.0.1")
    Process.sleep(400)
    Node.connect(:"node1@127.0.0.1")
    Process.sleep(400)
    Node.disconnect(:"node1@127.0.0.1")
end
```